Prevent from a PHP warning if the log file was deleted by an earlier killTask call

+++ packages/web/lib/service/multicasttask.class.php
@@ -696,7 +696,9 @@ class MulticastTask extends FOGService
     public function killTask()
     {
         $this->killTasking();
-        unlink($this->getUDPCastLogFile());
+        if (file_exists($this->getUDPCastLogFile())) {
+            unlink($this->getUDPCastLogFile());
+        }
         return true;
     }
     /**
